windows compat

+++ src/lib/random_color.cc
@@ -7,7 +7,7 @@ namespace tlz {
 	
 cv::Vec3b random_color(int i) {
 	std::mt19937 gen(i + 1);
-	std::uniform_int_distribution<uchar> dist(0, 255);
+	std::uniform_int_distribution<int> dist(0, 255);
 	return cv::Vec3b(dist(gen), dist(gen), dist(gen));
 }
